requirements.txt
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cssselect2==0.6.0
defusedxml==0.7.1
ftfy==6.1.1
idna==3.4
ipadic==1.0.0
langcodes==3.3.0
lxml>=4.9.2
mecab-python3==1.0.6
msgpack==1.0.4
orjson==3.8.0
Pillow>=9.2.0
plac==1.3.5
pycparser==2.21
regex==2022.10.31
requests==2.28.1
tinycss2==1.1.1
tqdm==4.64.1
unidic==1.1.0
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
wordfreq==3.0.3
setuptools>=68.2.2